Finish version 0.3
The version 0.3 is finished now. This version is for the major
revision to Bioinformatics. Basically we fixed several important
bugs in evalue estimation, so the results should change a lot.
However, the results of the first two experiments can be
guaranteed. Therefore, we don't have to worry a lot on the result.
We could publish this software to the website. Pls clean comments
before publishing the source code. The binaries should be built
in a clean environment. Below shows a rough description of the
improvement of version 0.3.

Bugfixs:
1) nan when evalue estimation is enabled.
2) true match count is ignored in computing evalue.
3) record order in multithread is not the same as in singlethread.

Refactor:
1) Use PPArray to replace the original generalized_peptides, to
   support the extension of mod.
2) Single-thread and multi-thread now share the same single
   spectrum search workflow.
3) Rename multiple files, classes and local variables.

Features:
1) New ModXcorr compute engine, optimize to 55% time.
2) Mod support in ModXcorr engine.
3) UI improvement, allow fix and mod specification from cmd.

In progress:
1) Variable modification enumeration (not tested).